==> src/03-getfollowers/03-lookupfollowers.Rmd <==
---
title: "GetBatchUsers"
author: "Johnatas Montezuma"
date: "10/11/2019"
output: html_document
---

We need to read the full followers list, split into sizable chunks of 4 hours and call getBatchUsers function. The script will save the batches in tmp folder and read the processed information.
Final output is a followers_summary data table with basic twitter information from users.

# Load followers_list
```{r}
followers_list <- readRDS(paste0(data_folder, "/followers_list.Rds"))
### followers_list <- top_n(followers_list, 2000000) ## select a shorter sample
```
Set as character for using getBatchUser function
```{r}
followers_list  <- followers_list %>%
  as.matrix() %>%  
  c() 
```


How long is the followers_list?
```{r}
prettyNum(length(followers_list), big.mark = ",")
```

split follower list in batches
```{r}
batch.size <- 180000 ### aprox 180K per hour, 2h
batches <- chunk(followers_list, n = batch.size)
rm(batch.size)
```

Save each batch from batch.size as single Rds file
```{r}
## loop for saving batches
for(i in 1:length(batches))
  {
    saveRDS(batches[[i]], file = paste0(data_folder, "/tmp/batch", i, ".Rds"))
}
```

Clean up
```{r}
### remove i and batches (as they are saved in tmp)
rm(i, batches)
```
                
Read files in tmp folder
```{r}
### read all files in tmp folder
batches_list <- list.files(paste0(data_folder,"/tmp"), full.names = TRUE, pattern = "batch")
```

For each batch, getUsersBatch and save as RDS
```{r}
### loop for each batch
for(i in 1:length(batches_list))
  {
        # set batch path
        batch.file <- batches_list[[i]] 
        # get batch number
        batch.number <- unlist(regmatches(x = batch.file, gregexpr("[0-9]+",text = batch.file))) 
        users <- readRDS(batch.file)       
        data <- getUsersBatch(ids = users, oauth = my_oauth)
        # save as Rds in processed folder
        saveRDS(data, file = paste0(data_folder, "/processed/batch", batch.number, "_data.Rds"))
        # move file to a done folder
        file.move(batch.file , paste0(data_folder, "/batches_done"))
}
```


clean up
```{r}
## clean up
rm(batches_list, i, users, data, batch.file, batch.number)
```



Read all processed folerand make  a list
```{r}
## read all files and make a full list
files <- list.files(path = paste0(data_folder, "/processed"), pattern ="_data.Rds", full.names = TRUE)
# this process takes a few minutes
followers_summary <- files %>%
        map(readRDS) %>% # read all files
        rbindlist() # combine all lists
```

Save as followers_summary in RDS in data
```{r}
saveRDS(followers_summary, paste0(data_folder, "/followers_summary.Rds"))
```

Clean up variables
```{r}
rm(files, followers.list, followers_summary)
```

```{r}
### remove files in tmp
files_to_delete <- list.files(paste0(data_folder, "/tmp"), full.names = TRUE) 
unlink(files_to_delete) ## deletes all files in tmp
files_to_delete <- list.files(paste0())
rm(files_to_delete)
```
