==> src/03-getfollowers/04-analyze-followers.Rmd <==
---
title: "Analyze followers"
author: "Johnatas Montezuma"
date: "10/11/2019"
output: html_document
---

Analyze all followers basic data from `followers_summary'

Read followers summary
```{r}
followers_summary <- readRDS(paste0(data_folder, "/followers_summary.Rds"))
```

Keep only unique data based on id_str
```{r}
# my_data %>% distinct(Sepal.Length, .keep_all = TRUE)
followers_summary <- followers_summary %>%
        distinct(id_str, .keep_all = TRUE)
```


Convert created at to a simple date format
```{r}
followers_summary <-  followers_summary %>%
        mutate(created_date = as.Date(followers_summary$created_at, format="%a %b %d %H:%M:%S %z %Y")) %>%
        mutate(created_at = NULL)
```

check users with top followers
```{r}
top_10_followers <- followers_summary %>% top_n(50, followers_count)
```

library(stringr)
top10_followers %>% 
    filter(str_detect(location, '\\b(River|Creek)\\b'))

followers_Peru<- followers_summary %>%
     filter(str_detect(location, "(Peru|Perú)"))
     
followers_Peru <- followers_Peru %>%
        top_n(1000, followers_count)
        

library(dplyr)
 df1 %>% 
     group_by(from) %>%
     filter(n_distinct(number)>1)


followers_listSample <- sample_n(followers_list, 500000)
## test selecting only that repeat 3 times
```{r}
TableSample1 <- followers_listSample %>%
    group_by(followers) %>%
    table()
```



## WIP ####
## scratch a sample
```{r}
followers_summaryTest <- head(followers_summary, 400)
followers_summaryTestDateFixed <-  followers_summaryTest %>%
        mutate(created_date = as.Date(followers_summaryTest$created_at, format="%a %b %d %H:%M:%S %z %Y")) %>%
        mutate(created_at = NULL)
        
```


## graph
## https://analyticsprofile.com/r-statistics-course/chapter-1-single-variable-analysis-in-r/
```{r}
library(beeswarm)
followers_count <- followers_summary$followers_count
beeswarm(followers_count)
```


boxplot(followers_count, horizontal=TRUE)

## top values
library(dplyr)
top10_final_subset2 = final_subset %>% group_by(u_id) %>% top_n(10,lift)
